+++ cobalt/media_session/media_session_test.cc
@@ -14,6 +14,7 @@
 
 #include "cobalt/media_session/media_session.h"
 
+#include <limits>
 #include <memory>
 
 #include "base/message_loop/message_loop.h"
@@ -48,9 +49,20 @@ class MockCallbackFunction : public MediaSession::MediaSessionActionHandler {
       Run, ReturnValue(const MediaSessionActionDetails& action_details));
 };
 
+class MockMediaSession : public MediaSession {
+ public:
+  explicit MockMediaSession(MediaSessionClient* client)
+      : MediaSession(client) {}
+  MOCK_CONST_METHOD0(GetMonotonicNow, SbTimeMonotonic());
+};
+
 class MockMediaSessionClient : public MediaSessionClient {
  public:
-  MOCK_METHOD0(OnMediaSessionChanged, void());
+  MockMediaSessionClient() : MediaSessionClient(new MockMediaSession(this)) {}
+  MockMediaSession& mock_session() {
+    return static_cast<MockMediaSession&>(*GetMediaSession().get());
+  }
+  MOCK_METHOD1(OnMediaSessionStateChanged, void(const MediaSessionState&));
 };
 
 MATCHER_P(SeekTime, time, "") {
@@ -71,58 +83,76 @@ TEST(MediaSessionTest, MediaSessionTest) {
 
   MockMediaSessionClient client;
 
-  ON_CALL(client, OnMediaSessionChanged())
+  ON_CALL(client, OnMediaSessionStateChanged(_))
       .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
-  EXPECT_CALL(client, OnMediaSessionChanged()).Times(1);
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(1);
 
   scoped_refptr<MediaSession> session = client.GetMediaSession();
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone,
+            client.GetMediaSessionState().actual_playback_state());
 
   session->set_playback_state(kMediaSessionPlaybackStatePlaying);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying,
+            client.GetMediaSessionState().actual_playback_state());
 
   run_loop.Run();
 }
 
-TEST(MediaSessionTest, GetActualPlaybackState) {
+TEST(MediaSessionTest, ActualPlaybackState) {
   base::MessageLoop message_loop(base::MessageLoop::TYPE_DEFAULT);
   base::RunLoop run_loop;
 
   MockMediaSessionClient client;
 
-  ON_CALL(client, OnMediaSessionChanged())
+  ON_CALL(client, OnMediaSessionStateChanged(_))
       .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
-  EXPECT_CALL(client, OnMediaSessionChanged()).Times(AtLeast(2));
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(AtLeast(2));
 
   scoped_refptr<MediaSession> session = client.GetMediaSession();
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone,
+            client.GetMediaSessionState().actual_playback_state());
 
   client.UpdatePlatformPlaybackState(kMediaSessionPlaybackStatePlaying);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying,
+            client.GetMediaSessionState().actual_playback_state());
 
   session->set_playback_state(kMediaSessionPlaybackStatePlaying);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying,
+            client.GetMediaSessionState().actual_playback_state());
 
   session->set_playback_state(kMediaSessionPlaybackStatePaused);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying,
+            client.GetMediaSessionState().actual_playback_state());
 
   client.UpdatePlatformPlaybackState(kMediaSessionPlaybackStatePaused);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePaused, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStatePaused,
+            client.GetMediaSessionState().actual_playback_state());
 
   session->set_playback_state(kMediaSessionPlaybackStateNone);
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone,
+            client.GetMediaSessionState().actual_playback_state());
 
   client.UpdatePlatformPlaybackState(kMediaSessionPlaybackStateNone);
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone,
+            client.GetMediaSessionState().actual_playback_state());
 
   run_loop.Run();
 }
@@ -132,15 +162,18 @@ TEST(MediaSessionTest, NullActionClears) {
   base::RunLoop run_loop;
 
   MockMediaSessionClient client;
+  MediaSessionState state;
 
-  ON_CALL(client, OnMediaSessionChanged())
+  ON_CALL(client, OnMediaSessionStateChanged(_))
       .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
-  EXPECT_CALL(client, OnMediaSessionChanged()).Times(AtLeast(0));
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(AtLeast(0));
 
   scoped_refptr<MediaSession> session = client.GetMediaSession();
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
-  EXPECT_EQ(0, client.GetAvailableActions().to_ulong());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone, state.actual_playback_state());
+  EXPECT_EQ(0, state.available_actions().to_ulong());
 
   MockCallbackFunction cf;
   EXPECT_CALL(cf, Run(_))
@@ -151,28 +184,33 @@ TEST(MediaSessionTest, NullActionClears) {
   FakeScriptValue<MediaSession::MediaSessionActionHandler> null_holder(NULL);
 
   session->SetActionHandler(kMediaSessionActionPlay, holder);
-  EXPECT_EQ(1, client.GetAvailableActions().to_ulong());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(1, client.GetMediaSessionState().available_actions().to_ulong());
   client.InvokeAction(kMediaSessionActionPlay);
 
   session->SetActionHandler(kMediaSessionActionPlay, null_holder);
-  EXPECT_EQ(0, client.GetAvailableActions().to_ulong());
+  base::RunLoop().RunUntilIdle();
+  EXPECT_EQ(0, client.GetMediaSessionState().available_actions().to_ulong());
   client.InvokeAction(kMediaSessionActionPlay);
 }
 
-TEST(MediaSessionTest, GetAvailableActions) {
+TEST(MediaSessionTest, AvailableActions) {
   base::MessageLoop message_loop(base::MessageLoop::TYPE_DEFAULT);
   base::RunLoop run_loop;
 
   MockMediaSessionClient client;
+  MediaSessionState state;
 
-  ON_CALL(client, OnMediaSessionChanged())
+  ON_CALL(client, OnMediaSessionStateChanged(_))
       .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
-  EXPECT_CALL(client, OnMediaSessionChanged()).Times(AtLeast(0));
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(AtLeast(0));
 
   scoped_refptr<MediaSession> session = client.GetMediaSession();
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
-  EXPECT_EQ(0, client.GetAvailableActions().to_ulong());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone, state.actual_playback_state());
+  EXPECT_EQ(0, state.available_actions().to_ulong());
 
   MockCallbackFunction cf;
   EXPECT_CALL(cf, Run(_)).Times(0);
@@ -180,66 +218,88 @@ TEST(MediaSessionTest, GetAvailableActions) {
 
   session->SetActionHandler(kMediaSessionActionPlay, holder);
 
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
   EXPECT_EQ(1 << kMediaSessionActionPlay,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->SetActionHandler(kMediaSessionActionPause, holder);
 
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
   EXPECT_EQ(1 << kMediaSessionActionPlay,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->SetActionHandler(kMediaSessionActionSeekto, holder);
 
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
   EXPECT_EQ(1 << kMediaSessionActionPlay | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   client.UpdatePlatformPlaybackState(kMediaSessionPlaybackStatePlaying);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPause | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->set_playback_state(kMediaSessionPlaybackStatePlaying);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPause | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->set_playback_state(kMediaSessionPlaybackStatePaused);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPause | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->set_playback_state(kMediaSessionPlaybackStatePlaying);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPause | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   client.UpdatePlatformPlaybackState(kMediaSessionPlaybackStatePaused);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePlaying, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPause | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->set_playback_state(kMediaSessionPlaybackStateNone);
 
-  EXPECT_EQ(kMediaSessionPlaybackStateNone, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStateNone, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPlay | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   session->set_playback_state(kMediaSessionPlaybackStatePaused);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePaused, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePaused, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPlay | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   client.UpdatePlatformPlaybackState(kMediaSessionPlaybackStateNone);
 
-  EXPECT_EQ(kMediaSessionPlaybackStatePaused, client.GetActualPlaybackState());
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(kMediaSessionPlaybackStatePaused, state.actual_playback_state());
   EXPECT_EQ(1 << kMediaSessionActionPlay | 1 << kMediaSessionActionSeekto,
-            client.GetAvailableActions().to_ulong());
+            state.available_actions().to_ulong());
 
   run_loop.Run();
 }
@@ -250,9 +310,9 @@ TEST(MediaSessionTest, SeekDetails) {
 
   MockMediaSessionClient client;
 
-  ON_CALL(client, OnMediaSessionChanged())
+  ON_CALL(client, OnMediaSessionStateChanged(_))
       .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
-  EXPECT_CALL(client, OnMediaSessionChanged()).Times(AtLeast(0));
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(AtLeast(0));
 
   scoped_refptr<MediaSession> session = client.GetMediaSession();
 
@@ -291,6 +351,156 @@ TEST(MediaSessionTest, SeekDetails) {
   details->set_action(kMediaSessionActionSeekbackward);
   details->set_seek_offset(5.6);
   client.InvokeAction(std::move(details));
+}
+
+TEST(MediaSessionTest, PositionState) {
+  base::MessageLoop message_loop(base::MessageLoop::TYPE_DEFAULT);
+  base::RunLoop run_loop;
+
+  MockMediaSessionClient client;
+
+  ON_CALL(client, OnMediaSessionStateChanged(_))
+      .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(AtLeast(3));
+
+  MockMediaSession& session = client.mock_session();
+  MediaSessionState state;
+
+  SbTimeMonotonic start_time = 1111111111;
+
+  base::Optional<MediaPositionState> position_state;
+  position_state.emplace();
+  position_state->set_duration(100.0);
+  position_state->set_position(10.0);
+
+  // Position state not yet reported
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(0,
+            state.GetCurrentPlaybackPosition(start_time + 999 * kSbTimeSecond));
+  EXPECT_EQ(0, state.duration());
+  EXPECT_EQ(0.0, state.actual_playback_rate());
+
+  // Forward playback
+  EXPECT_CALL(session, GetMonotonicNow()).WillOnce(Return(start_time));
+  position_state->set_playback_rate(1.0);
+  session.SetPositionState(position_state);
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ((10 + 50) * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 50 * kSbTimeSecond));
+  EXPECT_EQ(100 * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 150 * kSbTimeSecond));
+  EXPECT_EQ(100 * kSbTimeSecond, state.duration());
+  EXPECT_EQ(1.0, state.actual_playback_rate());
+
+  // Fast playback
+  EXPECT_CALL(session, GetMonotonicNow()).WillOnce(Return(start_time));
+  position_state->set_playback_rate(2.0);
+  session.SetPositionState(position_state);
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ((10 + 2 * 20) * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 20 * kSbTimeSecond));
+  EXPECT_EQ(100 * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 50 * kSbTimeSecond));
+  EXPECT_EQ(100 * kSbTimeSecond, state.duration());
+  EXPECT_EQ(2.0, state.actual_playback_rate());
+
+  // Reverse playback
+  EXPECT_CALL(session, GetMonotonicNow()).WillOnce(Return(start_time));
+  position_state->set_playback_rate(-1.0);
+  session.SetPositionState(position_state);
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(0 * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 20 * kSbTimeSecond));
+  EXPECT_EQ((10 - 3) * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 3 * kSbTimeSecond));
+  EXPECT_EQ(100 * kSbTimeSecond, state.duration());
+  EXPECT_EQ(-1.0, state.actual_playback_rate());
+
+  // Indefinite duration (live) playback
+  EXPECT_CALL(session, GetMonotonicNow()).WillOnce(Return(start_time));
+  position_state->set_duration(std::numeric_limits<double>::infinity());
+  position_state->set_playback_rate(1.0);
+  session.SetPositionState(position_state);
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(10 * kSbTimeSecond + 1 * kSbTimeDay,
+            state.GetCurrentPlaybackPosition(start_time + 1 * kSbTimeDay));
+  EXPECT_EQ(kSbTimeMax, state.duration());
+  EXPECT_EQ(1.0, state.actual_playback_rate());
+
+  // Paused playback
+  // (Actual playback rate is 0.0, so position is the last reported position.
+  //  The web app should update position and playback states together.)
+  session.set_playback_state(kMediaSessionPlaybackStatePaused);
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(10 * kSbTimeSecond,
+            state.GetCurrentPlaybackPosition(start_time + 999 * kSbTimeSecond));
+  EXPECT_EQ(kSbTimeMax, state.duration());
+  EXPECT_EQ(0.0, state.actual_playback_rate());
+  session.set_playback_state(kMediaSessionPlaybackStatePlaying);
+
+  // Position state cleared
+  EXPECT_CALL(session, GetMonotonicNow()).WillOnce(Return(start_time));
+  session.SetPositionState(base::nullopt);
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  EXPECT_EQ(0,
+            state.GetCurrentPlaybackPosition(start_time + 999 * kSbTimeSecond));
+  EXPECT_EQ(0, state.duration());
+  EXPECT_EQ(0.0, state.actual_playback_rate());
+}
+
+TEST(MediaSessionTest, Metadata) {
+  base::MessageLoop message_loop(base::MessageLoop::TYPE_DEFAULT);
+  base::RunLoop run_loop;
+
+  MockMediaSessionClient client;
+
+  ON_CALL(client, OnMediaSessionStateChanged(_))
+      .WillByDefault(InvokeWithoutArgs(&run_loop, &base::RunLoop::Quit));
+  EXPECT_CALL(client, OnMediaSessionStateChanged(_)).Times(AtLeast(0));
+
+  MockMediaSession& session = client.mock_session();
+  MediaSessionState state;
+
+  MediaMetadataInit init_metadata;
+  base::Optional<MediaMetadataInit> state_metadata;
+
+  // Metadata not yet set
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  state_metadata = state.metadata();
+  EXPECT_FALSE(state.has_metadata());
+  EXPECT_FALSE(state_metadata.has_value());
+
+  // Set metadata and make sure it gets into the MediaSessionState
+  init_metadata = MediaMetadataInit();
+  init_metadata.set_title("title");
+  init_metadata.set_artist("artist");
+  init_metadata.set_album("album");
+  MediaImage art_image;
+  art_image.set_src("http://art.image");
+  script::Sequence<MediaImage> artwork;
+  artwork.push_back(art_image);
+  init_metadata.set_artwork(artwork);
+  session.set_metadata(
+      scoped_refptr<MediaMetadata>(new MediaMetadata(init_metadata)));
+
+  base::RunLoop().RunUntilIdle();
+  state = client.GetMediaSessionState();
+  state_metadata = state.metadata();
+  EXPECT_TRUE(state.has_metadata());
+  EXPECT_TRUE(state_metadata.has_value());
+  EXPECT_EQ("title", state_metadata->title());
+  EXPECT_EQ("artist", state_metadata->artist());
+  EXPECT_EQ("album", state_metadata->album());
+  EXPECT_EQ(1, state_metadata->artwork().size());
+  EXPECT_EQ("http://art.image", state_metadata->artwork().at(0).src());
 }
 
 }  // namespace